move view buttons to a resusable component
