+++ orca/server/static/js/table.js
@@ -1,52 +1,9 @@
 var React = require('react');
-var classNames = require('classnames');
 var _ = require('lodash');
+
 var dtopts = require('./dtopts.js');
+var ViewButtonGroup = require('./buttons.js');
 
-
-var TableButton = React.createClass({
-  clickHandler: function(event) {
-    this.props.clickHandler(this.props.view);
-  },
-  render: function() {
-    var classes = classNames({
-      "btn": true,
-      "btn-default": true,
-      "active": this.props.active
-    });
-
-    return (
-      <button
-          type="button"
-          className={classes}
-          onClick={this.clickHandler}>
-        {this.props.text}
-      </button>
-    );
-  }
-});
-
-var TableButtons = React.createClass({
-  render: function() {
-    var button_els = _.map(this.props.buttons, function(b) {
-      return (
-        <TableButton
-          text={b.text}
-          view={b.view}
-          key={b.view}
-          active={b.view === this.props.view}
-          clickHandler={this.props.clickHandler}
-        />
-      );
-    }.bind(this));
-
-    return (
-      <div className="tableButtons btn-group" role="group">
-        {button_els}
-      </div>
-    );
-  }
-});
 
 var TablePreview = React.createClass({
   componentDidMount: function() {
@@ -185,7 +142,7 @@ var TableApp = React.createClass({
         <div className="page-header">
           <h1>{this.props.table}</h1>
         </div>
-        <TableButtons
+        <ViewButtonGroup
             buttons={this.tableButtons}
             view={this.state.view}
             clickHandler={this.handleButtonClick} />
